Add lang en_short, for e.g. "4w 2d ago" shorthand.
I'm not entirely sure this is the best implementation of this, but it's the
simplest, and I don't even know if you'll want to accept the feature.

The other way that might be better would be for each defined language to provide
both long and short forms of each string, and let the constructor take an
additional "terse" or similarly-named option which tells it that you want the
short-form output rather than longer.

Thoughts welcome?